broke apart the cats view into several different functions to clean up
