+++ fosteriteapp/views/cat/__init__.py
@@ -1,2 +1,5 @@
 from .cat_serializer import CatSerializer
 from .cats import Cats
+from .cat_create import cat_create
+from .cat_list import cat_list
+from .cat_retrieve import cat_retrieve
